Code refactorization and best practices.
