+++ index.html
@@ -193,10 +193,9 @@
       </footer>
     </section>
 
-    <img class="mypicture" src="./images/myself.png" alt="" />
-
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.11.0/gsap.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.11.0/ScrollTrigger.min.js"></script>
+
 
     <script src="./js/popup.js" type="module"></script>
     <script src="./js/script.js" type="module"></script>
